+++ enclone_visual/src/process_messages.rs
@@ -12,7 +12,6 @@ use flate2::read::GzDecoder;
 use gui_structures::ComputeState::*;
 use iced::{Clipboard, Color, Command};
 use io_utils::*;
-use std::fs::File;
 use std::io::Read;
 use std::time::{Duration, Instant};
 use vector_utils::*;
@@ -123,63 +122,7 @@ impl EncloneVisual {
                 Command::none()
             }
 
-            Message::DoShare(check_val) => {
-                self.do_share = check_val;
-                if !check_val {
-                    self.do_share_complete = false;
-                } else {
-                    let mut recipients = Vec::<String>::new();
-                    for i in 0..self.user_value.len() {
-                        if self.user_valid[i] {
-                            recipients.push(self.user_value[i].clone());
-                        }
-                    }
-                    let mut index = 0;
-                    for i in 0..self.archive_share_requested.len() {
-                        if self.archive_share_requested[i] {
-                            index = i;
-                        }
-                    }
-                    let path = format!(
-                        "{}/{}",
-                        self.archive_dir.as_ref().unwrap(),
-                        self.archive_list[index]
-                    );
-                    if !path_exists(&path) {
-                        xprintln!("could not find path for archive file\n");
-                        std::process::exit(1);
-                    }
-                    let mut content = Vec::<u8>::new();
-                    let f = File::open(&path);
-                    if f.is_err() {
-                        xprintln!("could not open archive file\n");
-                        std::process::exit(1);
-                    }
-                    let mut f = f.unwrap();
-                    let res = f.read_to_end(&mut content);
-                    if res.is_err() {
-                        xprintln!("could not read archive file\n");
-                        std::process::exit(1);
-                    }
-                    SHARE_CONTENT.lock().unwrap().clear();
-                    SHARE_CONTENT.lock().unwrap().push(content);
-                    SHARE_RECIPIENTS.lock().unwrap().clear();
-                    let days = Utc::now().num_days_from_ce();
-                    for i in 0..recipients.len() {
-                        SHARE_RECIPIENTS.lock().unwrap().push(recipients[i].clone());
-                        let mut user_name = [0 as u8; 32];
-                        for j in 0..recipients[i].len() {
-                            user_name[j] = recipients[i].as_bytes()[j];
-                        }
-                        self.shares.push(Share {
-                            days_since_ce: days,
-                            user_id: user_name,
-                        });
-                    }
-                    SENDING_SHARE.store(true, SeqCst);
-                }
-                Command::perform(compute_share(), Message::CompleteDoShare)
-            }
+            Message::DoShare(check_val) => do_share_button_pressed(self, check_val),
 
             Message::CompleteDoShare(_) => {
                 self.do_share_complete = true;
